Move permission code to indico.core.permissions
